requirements.txt
colorama==0.4.6
prompt_toolkit==3.0.39
pandas==2.1.1
dnspython==2.4.2
impacket==0.11.0